+++ demo/webpack/index.js
@@ -1,6 +1,6 @@
 var j = require("./node_modules/jsplumb/dist/js/jsplumb.js").jsPlumb.getInstance({
     Connector: "Flowchart",
-    Anchor: "Bottom",
+    anchor: "Bottom",
     Endpoint: [ "Dot", { radius: 2 }],
     ConnectionOverlays: [
         [ "Arrow", { location: 0, width: 10, length: 7, foldbackPoint: 0.62, direction:-1 }]